+++ pwctl/__init__.py
@@ -1,3 +1,3 @@
 """PipeWire Controller."""
 
-__version__ = '0.3.4'
+__version__ = '0.3.5'